Добавляет правки в сообщениях об ошибке
