+++ js/main.js
@@ -1,3 +1,2 @@
 import './user-pictures.js';
 import './form-validation.js';
-import './user-modal.js';
