+++ js/user-modal.js
@@ -5,6 +5,11 @@ const uploadPictureButton = userPictureForm.querySelector('.img-upload__input');
 const pictureEditorModal = userPictureForm.querySelector('.img-upload__overlay');
 const editorCloseButton = userPictureForm.querySelector('.img-upload__cancel');
 
+const successModal = document.querySelector('#success').content.querySelector('.success');
+const successButton = successModal.querySelector('.success__button');
+const errorModal = document.querySelector('#error').content.querySelector('.error');
+const errorButton = errorModal.querySelector('.error__button');
+
 function closeUserModal () {
   pictureEditorModal.classList.add('hidden');
   document.body.classList.remove('modal-open');
@@ -12,9 +17,7 @@ function closeUserModal () {
   document.removeEventListener('keydown', onModalEscKeydown);
 }
 
-editorCloseButton.addEventListener('click', () => {
-  closeUserModal();
-});
+editorCloseButton.addEventListener('click', closeUserModal);
 
 function onModalEscKeydown (evt) {
   if (isEscapeKey(evt)) {
@@ -29,8 +32,36 @@ function openUserModal () {
   document.addEventListener('keydown', onModalEscKeydown);
 }
 
-uploadPictureButton.addEventListener('change', () => {
-  openUserModal();
+uploadPictureButton.addEventListener('change', openUserModal);
+
+//Окно с ошибкой или удачной отправкой формы
+
+const showErrorModal = function () {
+  document.body.appendChild(errorModal);
+};
+
+const showSuccessModal = function () {
+  document.body.appendChild(successModal);
+};
+
+const hideErrorModal = function () {
+  errorModal.remove();
+};
+
+errorButton.addEventListener('click', hideErrorModal);
+
+document.addEventListener('keydown', () => {
+  if(isEscapeKey || errorModal) {
+    hideErrorModal();
+  }
 });
 
-export { userPictureForm ,closeUserModal };
+const hideSuccessModal = function () {
+  successModal.remove();
+  userPictureForm.submit();
+  closeUserModal();
+};
+
+successButton.addEventListener('click', hideSuccessModal);
+
+export { userPictureForm, showErrorModal, showSuccessModal };
